requirements.txt
djangorestframework==3.11.0
Django==2.1.15
django-cors-headers==3.2.1